move all addresses to helper
